Create CodeNotWorking
My goal this week was to direct logins to the correct user's page.  I spent a lot of time on it, and didn't finish.  I will be meeting Amanda during her office hours tomorrow to attempt to fix it.  I also didn't commit or push this code because my teammates are having issues with GitHub.  We aren't sure if we will lose everything if I pull before I push.  Hopefully we will have this fixed before class tomorrow.